Look up episodes on TVmaze instead of TheTVDB
TheTVDB's version 1 API is retired. Its search endpoint answers every
query with an empty document, so no show ever resolved and nothing could
be renamed. TVmaze serves the same data as JSON and needs no API key.

TVmaze publishes only the aired ordering, so the DVD ordering preference
now always falls back to the aired placement. It was never exposed in the
UI. The tests that covered it, and the ~60 tests disabled behind
RUN_EXTRA_TESTS, all encoded TheTVDB's identifiers and orderings, so they
go with it.
