+++ src/main/java/org/tvrenamer/model/ShowStore.java
@@ -1,7 +1,7 @@
 package org.tvrenamer.model;
 
 import org.tvrenamer.controller.ShowInformationListener;
-import org.tvrenamer.controller.TheTVDBProvider;
+import org.tvrenamer.controller.TVmazeProvider;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
@@ -229,7 +229,7 @@ public class ShowStore {
         Callable<Boolean> showFetcher = () -> {
             ShowOption showOption;
             try {
-                TheTVDBProvider.getShowOptions(showName);
+                TVmazeProvider.getShowOptions(showName);
                 showOption = showName.selectShowOption();
             } catch (DiscontinuedApiException e) {
                 showName.apiDiscontinued();
